+++ test/src/test/java/de/njsm/tinyauth/test/ApiTests.java
@@ -43,7 +43,7 @@ import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.equalToIgnoringCase;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-public class ApiTests implements TinyAuthTest {
+public class ApiTests implements TinyAuthTest, ApiGadgets {
 
     private Client client = Clients.getConfidentialClient();
 
@@ -69,7 +69,7 @@ public class ApiTests implements TinyAuthTest {
 
         assertEquals(Set.of("openid"), Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
-        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, client);
+        tokenAsserter().verifyAccessToken(tokenResponse.getString(ID_TOKEN), client, client);
     }
 
     @Test
@@ -82,7 +82,7 @@ public class ApiTests implements TinyAuthTest {
 
         assertEquals(scopes, Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
-        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, user);
+        tokenAsserter().verifyAccessToken(tokenResponse.getString(ID_TOKEN), client, user);
     }
 
     @Test
@@ -98,7 +98,7 @@ public class ApiTests implements TinyAuthTest {
 
         assertEquals(Set.of("openid"), Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
-        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, user);
+        tokenAsserter().verifyAccessToken(tokenResponse.getString(ID_TOKEN), client, user);
     }
 
     @Test
@@ -144,7 +144,7 @@ public class ApiTests implements TinyAuthTest {
 
         assertEquals(scopes, Set.of(tokenResponse.getString(SCOPE).split(" ")));
         assertEquals(tokenResponse.getString(ACCESS_TOKEN), tokenResponse.getString(ID_TOKEN), "access token different from id token");
-        tokenAsserter().verifyAccessTokenWithoutNonce(tokenResponse.getString(ID_TOKEN), client, client);
+        tokenAsserter().verifyAccessToken(tokenResponse.getString(ID_TOKEN), client, client);
     }
 
     private String buildTokenFromSharedSecret(Client client) throws Exception {