+++ test/src/test/java/de/njsm/tinyauth/test/oidc/TokenAsserter.java
@@ -26,8 +26,8 @@ import com.nimbusds.jose.proc.SecurityContext;
 import com.nimbusds.jwt.JWTClaimsSet;
 import com.nimbusds.jwt.proc.ConfigurableJWTProcessor;
 import com.nimbusds.jwt.proc.DefaultJWTProcessor;
-import de.njsm.tinyauth.test.data.OidcToken;
 import de.njsm.tinyauth.test.data.Client;
+import de.njsm.tinyauth.test.data.OidcToken;
 import de.njsm.tinyauth.test.data.User;
 import de.njsm.tinyauth.test.repository.Endpoints;
 import io.restassured.path.json.JsonPath;
@@ -41,33 +41,12 @@ import static java.util.Collections.singletonList;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class TokenAsserter {
+public interface TokenAsserter {
+    OidcToken verifyAccessToken(String token, Client client, User user) throws Exception;
 
-    public OidcToken verifyAccessTokenWithoutNonce(String token, Client client, User user) throws Exception {
-        JWTClaimsSet claims = verifyToken(token);
-        verifyAccessTokenClaimsWithoutNonce(claims.getClaims(), client, user);
-        return new OidcToken(token, claims);
-    }
+    OidcToken verifyRefreshToken(String token, Client client, User user, Set<String> scopes) throws Exception;
 
-    public OidcToken verifyAccessToken(String token, Client client, User user, String nonce) throws Exception {
-        JWTClaimsSet claims = verifyToken(token);
-        verifyAccessTokenClaims(claims.getClaims(), client, user, nonce);
-        return new OidcToken(token, claims);
-    }
-
-    public OidcToken verifyRefreshToken(String token, Client client, User user, String nonce, Set<String> scopes) throws Exception {
-        JWTClaimsSet claims = verifyRefreshTokenSpecificClaims(token, scopes);
-        verifyAccessTokenClaims(claims.getJSONObjectClaim(ACCESS_TOKEN), client, user, nonce);
-        return new OidcToken(token, claims);
-    }
-
-    public OidcToken verifyRefreshTokenWithoutNonce(String token, Client client, User user, Set<String> scopes) throws Exception {
-        JWTClaimsSet claims = verifyRefreshTokenSpecificClaims(token, scopes);
-        verifyAccessTokenClaimsWithoutNonce(claims.getJSONObjectClaim(ACCESS_TOKEN), client, user);
-        return new OidcToken(token, claims);
-    }
-
-    public void verifyUserinfo(JsonPath userinfo, JWTClaimsSet accessTokenClaims) {
+    default void verifyUserinfo(JsonPath userinfo, JWTClaimsSet accessTokenClaims) {
         Map<String, Object> convertedClaims = new HashMap<>(accessTokenClaims.getClaims());
         convertedClaims.put(EXPIRATION_TIME, ((Date) convertedClaims.get(EXPIRATION_TIME)).getTime() / 1000);
         convertedClaims.put(ISSUANCE_TIME, ((Date) convertedClaims.get(ISSUANCE_TIME)).getTime() / 1000);
@@ -88,22 +67,16 @@ public class TokenAsserter {
         assertEquals(convertedClaims, convertedUserinfo);
     }
 
-    private JWTClaimsSet verifyRefreshTokenSpecificClaims(String token, Set<String> scopes) throws Exception {
-        JWTClaimsSet claims = verifyToken(token);
-
-        assertEquals(Endpoints.getIssuer(), claims.getStringClaim(ISSUER));
-        assertEquals(scopes, new HashSet<>(claims.getStringListClaim(SCOPES)));
-        assertTrue(claims.getExpirationTime().after(new Date()), "token has already expired");
-
-        return claims;
+    default JWTClaimsSet verifyToken(String token) throws Exception {
+        ConfigurableJWTProcessor<SecurityContext> jwtProcessor = new DefaultJWTProcessor<>();
+        JWKSource<SecurityContext> keySource = new RemoteJWKSet<>(new URL(Endpoints.getJwksUrl()));
+        JWSAlgorithm expectedJWSAlg = JWSAlgorithm.ES384;
+        JWSKeySelector<SecurityContext> keySelector = new JWSVerificationKeySelector<>(expectedJWSAlg, keySource);
+        jwtProcessor.setJWSKeySelector(keySelector);
+        return jwtProcessor.process(token, null);
     }
 
-    private void verifyAccessTokenClaims(Map<String, Object> claims, Client client, User user, String nonce) {
-        verifyAccessTokenClaimsWithoutNonce(claims, client, user);
-        assertEquals(nonce, claims.get(NONCE));
-    }
-
-    private void verifyAccessTokenClaimsWithoutNonce(Map<String, Object> claims, Client client, User user) {
+    default void verifyAccessTokenClaims(Map<String, Object> claims, Client client, User user) {
         Date now = new Date();
 
         assertEquals(Endpoints.getIssuer(), claims.get(ISSUER));
@@ -125,19 +98,20 @@ public class TokenAsserter {
         assertTrue(authTime.compareTo(issuanceTime) < 1, "token was issued (" + issuanceTime + ") before authentication (" + authTime + ")");
     }
 
-    private Date castToDate(Object rawDate) {
+    default Date castToDate(Object rawDate) {
         if (rawDate instanceof Date)
             return (Date) rawDate;
         else
             return new Date((Long) rawDate * 1000);
     }
 
-    private static JWTClaimsSet verifyToken(String token) throws Exception {
-        ConfigurableJWTProcessor<SecurityContext> jwtProcessor = new DefaultJWTProcessor<>();
-        JWKSource<SecurityContext> keySource = new RemoteJWKSet<>(new URL(Endpoints.getJwksUrl()));
-        JWSAlgorithm expectedJWSAlg = JWSAlgorithm.ES384;
-        JWSKeySelector<SecurityContext> keySelector = new JWSVerificationKeySelector<>(expectedJWSAlg, keySource);
-        jwtProcessor.setJWSKeySelector(keySelector);
-        return jwtProcessor.process(token, null);
+    default JWTClaimsSet verifyRefreshTokenSpecificClaims(String token, Set<String> scopes) throws Exception {
+        JWTClaimsSet claims = verifyToken(token);
+
+        assertEquals(Endpoints.getIssuer(), claims.getStringClaim(ISSUER));
+        assertEquals(scopes, new HashSet<>(claims.getStringListClaim(SCOPES)));
+        assertTrue(claims.getExpirationTime().after(new Date()), "token has already expired");
+
+        return claims;
     }
 }